Create creature parser
This detect h1 elements inside the div.body tag, which hold the creature
name, but more importantly represent the start of an individual
creature. It scans each element until it reaches another H1 or the end
of the document. It finally returns all of the tags related to that
monster in a Nokogiri::HTML::DocumentFragment