Improve parent element detection for senses parser
The way I detect where senses are, is I find each <b> or <strong>
element that contains the text "Senses", then I return is parent element
and start the parsing from that parent elements text.

Before this change, I was only look at <b> and not <strong>. Now it
looks for both.