Add parent_element to Type parser
This will return the element that contains the type and subtype for
easier parsing of the actual types. It will also allow me to parse
multiple subtypes because I will have a single element to group all of
the subtypes. Otherwise I would have to search every element for every
subtype, or something.