Add Armor attribute
This will store any type of armor bonus; ac, flat-footed, dex, size,
etc.

The idea is I will store each of these as a row in a table that links
back to a monster. For example:

Monsters
id | title | ...
--   -----
1    Wolf
2    Bat

Armor
id | title       | bonus | monster_id
--   -----------   -----   ----------
1    ac            10      1
2    touch         7       1
3    flat_footed   7       1
4    ac            12      2
5    ...

I thought maybe I should do ac, touch, and flat_footed on the armor or
monster itself, but this strategy will result in less joins. Most
monsters have _some_ extra bonus like Dex or size. If I did an Armor
table for the main three attributes, it'd have to join the extra
attributes to that table, and join armor back to the monster. This is
more uniform.