Add armor parser
This takes an element like this:
<p class="stat-block-1"><b>AC</b> 21, touch 17, flat-footed 13 (+7 Dex,
+1 dodge, +4 natural, –1 size)</p>

And parses it into an array of Armor attributes like this:

title: 'ac', bonus: 21
title: 'touch', bonus: 17
title: 'dex', bonus: 7
title: 'size', bonus: -1
etc.

The way it works is by first finding the above element. It looks at all
the stats and finds the first element starting with "AC ". Then it
sanitizes the text to be lowercase, converts " (" and "; " into commas,
removes
")", and converts en dashes "–" into hyphens "-".

Then it has a string like: "ac 21, touch 17, flat-footed 13, +7 dex, +1
dodge, +4 natural, –1 size". It splits the string by each comma so it
has
pairs like: "ac 21". Each pair then is scanned for digits "21", "+7",
etc.,
for the bonus and then the bonus is removed from the text, resulting in
the title.

Each value then is used to initialize a new Armor.