Add Dice parser
A parser for multiple dice in a string. For now, they are only tested
to be together in a string like this: " 2 claws +18 (2d8/19–20 plus
fear and grab)" or "hp 44 (3 HD; 2d8+1d10+3)"

It works by putting a StringScanner in a while loop, scanning a string
for dice. If it finds one, it will add it to an array. If it can't find
one, it will look for a bonus like "+11" and add that to the last die
parsed. If it can't find either, it will skip ahead one character so
that the loop doesn't run forever.

Once it has strings, it splits up each part into the count, sides, and
bias of the dice. 1d10+3 => "1", "d10", "+3". It then converts each
value into an integer and creates a new Die model. Finally, it returns
all the dice in an array.