Add DamageReduction parser
The parser works like most others, by finding the enclosing parent
element that contains the DamageReduction if there is one. I've
improved on that a little though by stopping the lookup after I find
the OFFENSE line. By that point in the scan, I'm passed any defensive
attributes and don't need or want to look any farther. Sometimes there
are other bits of the text that have DR and I don't want to scan those.

After finding the parent element, I use a StringScanner to scan for
text between DR and whatever may come after DR including other
defensive attributes, a semicolon, the end of the line, or even another
DR (see Vampire Savage).

Then it's a matter of extracting the amount (some digit(s)) and the
exceptions if there are any. Most times if there are no exceptions, an
em- or en-dash is used. I normalize those and strip that as necessary.
Splitting up the exceptions isn't too bad on this one because I've only
found "and" or "or" separations.

Anyway, then it returns a DamageReduction attribute which is just a
subclass of DefenseWithExceptions with literally nothing added.