Add Spell Resistance (SR) parser
This is pretty simple. It detects the parent element like other
parsers, then extract a digit (or 2) beyond the text SR.

The only thing this doesn't cover is extra text that applies to SR. For
example, the Thulgant Qlippoth has "SR 25 vs. lawful spells and
creatures". This parser doesn't care and only reports 25.

This is because there are so few SR's with extra text (I scanned
everything and found 3) that I can't imagine searching for those
attributes.