Fix bug in creature parser
The creature parser wasn't doing its job correctly. It was searching
for h1/h2's in the document and using those as an anchor point to scan
the rest for creatures.

Now it scans all paragraphs for some amount of text plus CR X where X
is one or more digits. This is a more accurate starting point for
creatures. Combine this with also checking for a minimum number of
stats and it seems pretty reliable. We'll see if it holds up.