Parse title from speed text
A title is always a single word ("movement", "fly", "burrow") that
comes before feet. For example "Speed 30 ft.", "fly 60 ft. (perfect)".
In the case of "Speed" which always exists as the start of the line, or
no word in front of the feet, we return "movement" which is the default
title.

I went through a few iterations of this, but it turns out the simplest
is the best. The parser searches for text right before a space and a
digit, then cleans it up afterward. That is the most accurate way (so
far) of getting the Speed's title.

This is instead of putting all the logic in the regular expression
which was very error prone. Now I just let the simple regex do its
thing, and return the default "movement" at certain points to avoid
errors.