Add armor_indexes to speed parser
Armor indexes indicate if a particular speed is "in armor" or not. Such
as Speed 30 ft. (20 ft. in armor)

It expects the speeds that are broken up from the divide method. Armor
speed are always listed after normal speeds and there are always the
same number as the non-armor speeds, such as 2 without armor and 2 with
armor. Occasionally there is a speed without armor _after_ the armor is
specified.

It first finds the index of the speed that specifies armor (the last
armored speed), divides it by two and rounds down. Now it has the last
"without" and "with" armor indexes. Then, iterates through each armor
and checks the range. If it's before the armor starts, it marks it as
false. If it's between that and the max armor index, it marks it as
true. After that, it marks everything false.

Example:

Speed 40 ft., fly 60 ft. (good); 30 ft., fly 40 ft. (good) in armor

The first two (40 ft. and fly 60 ft.) are without armor, and the last
two (30 ft. and fly 40 ft.) are in armor.