Add title method to Attack parser
This uses a shared StringScanner to scan past any possible attack count
and scan just for the text before bonuses.

First, it skips an initial digit and space if there is one. It then
looks for a space with a + or - which is the start of a bonus (and
never the start of an attack because of the starting space). Then
subtract what it found from what is scanned over.

Example:

text = +1 dagger +13/+8 (1d4+4/19–20)
* skip initial digit and space (no change)
* scan until " +" which captures "+1 dagger +"
* remove match (" +") from captured text leaving "+1 dagger"