Add bonuses method to attack parser
This is a little more complicated. First, here's what it's trying to
parse (input):

* +2 wounding spear +32/+27/+22/+17 (3d6+17/x3 plus 1 bleed)
* incorporeal touch +6 (1d6 negative energy plus 1d6 Con drain)
* short sword –1 (1d3–3/19–20)

And here's the desired output:

* [32, 27, 22, 17]
* [6]
* [-1]

First, the scanner skip a bunch of stuff we don't care about like the
initial attack count, or a bonus to the weapon. These can both
interfere with the matching of bonuses because they look very similar.

Then it scans for a positive or negative number like +4, -1, etc. and
adds it to the an array. After that it puts the scanner into a loop
looking for a slash, then a positive or negative number again, over and
over until it can't find anymore slashes.

Finally, it splits up the found text by its slashes and maps them to
integers.