Add damage method to attack parser
The work for this was mostly done by the Dice parser already. All it
needs to do is use a combination of the regexes found in the Dice
parser to find the string (digits "d" digits, then an optional "+"
digits), then send it to the parser. It returns an array of dice so I
just grab the first one (which is the only one).

Occasionally there is no specific damage roll, like with boggards
(monsterCodex/boggards.html). In these cases, we want to return nil for
the damage because it's not parseable within the line. It either does
no damage, or the information is a special ability defined elsewhere.