Add critical range method to attack parser
This uses the fact that every attack has exactly one dice roll as a
starting point. The scanner scans to the die, then looks for a trailing
slash. If there's no trailing slash, it returns the default of 1. This
is an indicator that there is either a critical range or a critical
multiplier.

The scanner looks for something that looks like a critical range. If it
doesn't find anything, it again returns the default of 1. Then it
splits up the string it's found by the `-` character and converts them
to integers. It makes a range from those integers. The size of that
range is the critical range.

Note: The top of the range is always 20. The parser could have not
split the text by `-` and just converted it to an integer. Text like
"19-20".to_i returns `19`. That could have then been fed into the range
with a fixed 20 at the top. However, while this is less code, it relies
on an edge case of `to_i` and a magic number so I decided against it.