Add additional effects method to attack parser
This extracts other effects of the attack like electricity or blindness:

* 2 slams +27 (2d6+12 plus 3d6 electricity or sonic and blindness or
deafness)
* torturous touch +11 touch (2d6 plus 1d6 Dex damage and convulsions)

The scanner starts by skipping until the parenthesis. This is so we
avoid any text before the effects. Then, it skips the dice roll if it
exists. Finally, it skips "plus " (including that space).

After that, it scans for word and space characters to get the
additional effects. if it doesn't find any (i.e. if none exist) it
returns an empty array. If it does find them, it will split the text by
"and" and "or", finally stripping them of extra space.