Fix edge case for attack parser's additional effects
When scanning for additional effects, the StringScanner skips to the
open parenthesis, skips the die, then would skip until the " plus ",
then scan for text and spaces. When an attack had no additional effects
but critical modifiers (a range or multiplier) it would get past the
die, try to skip the plus (and return nil) then scan for text and
spaces.

At this point the StringScanner would be at the start of the critical
info and pick up text like "/15" of "1d8+18/15-20" which is not an
additional effect. Now, the scanner skips that critical info also.