Fix Attack parser for swarm attacks
Swarm attacks look like this: "swarm (3d6)"

Finding "title" required that the scanner know where to stop. This used
to look for the start of a bonus (+1/+2) and get the text before that.
However, in a swarm attack, there is no explicit bonus. so now it looks
for either the bonuses or an open parenthesis.