Start melee parser with parent_text method
parent_text is similar to the familiar parent_element, but returns text
instead of an element. More importantly, it can scan over multiple
elements to find what it needs.

It does this by starting with the parent element, and grabbing
successive siblings. It checks every sibling for:

* nil-ness (end of the list of siblings)
* being a text-type element
* having any <strong> or <b> elements (start of a new attribute)
* other section headers (STATISTICS, TACTICS, etc)

If it finds any of these it breaks the loop or skips this particular
element. If the sibling passes all these tests, it adds its text to the
original parent's text.