Parse attack strings into attack attributes
This uses a StringScanner on the text that parent_text returns. It's
technique is to scan for close parenthesis, check to make sure it's not
in the middle of an open parenthesis, then add the text it found to the
array of strings.

Additionally, it skips joining text like commas, "or", "and", etc. so
the Attack parser doesn't have to deal with extra text. It also skips
the "Melee " text at the front.