Refactor additional effects of attack parser
This has been causing me a bunch of problems, so I refactored the
method.

Since additional effects are all at the end, reversing then scanning
through the string is helpful.

It starts with a StringScanner from a reversed string. Then:

* Skips a closed parenthesis to get past what is often the end of the
string (but not always)
* Scans until it finds "plus " to grab the most common type of
additional effects like: "plus stun and grab"
* If that didn't find anything, it's possible the text has no "plus" or
anything. At first I tried to search for the open parenthesis, but that
was foley. Too much garbage to sort out there. So instead, I look for
multiple alpha and space characters starting wherever it scanned, which
at this point is from the very end or just after a closing parenthesis.
* If it _still_ hasn't found anything after that, I assume there are no
additional effects and return an empty array.

Then I filter out things like "plus ", re-reverse the string back,
split by "and", and "or", and strip out extra whitespace.