Slight refactor of melee parser
This moves the attack scanning out into its own method. Also it now
scans only for a close parenthesis at first, and not the end of the
line because it's better about rescanning the same line to catch
end-of-line content when there are no more attacks left in the string.