Parse feet from the text of the parent element
Nothing to crazy here. Occasionally there is a number like "2-1/2 ft."
I deal with that by simply replacing the text "-1/2" with ".5", then
converting it to floating point.

I'm also using the fact that calling `to_f` or `to_i` on a string will
start parsing the start of the string and stop when it finds a
non-numeric character. It ends up parsing things like "15 ft." into
`15`. Earlier tries at this code tried to remove " ft." as well, but it
seemed unnecessary and added more complexity.