Add Special Attacks parser
This is similar to the other parsers, but let's walk through it anyway.

Here are some examples of special attacks:

* Special Attacks gaze
* Special Attacks spit acid
* Special Attacks distraction (DC 11), wounding
* Special Attacks constrict (tail slap, 2d6+10 plus crushing coils),
infuse weapon, multiweapon mastery
* Special Attacks +1 on attack rolls against goblinoid and orc humanoids

I decided to skip anything inside of the parenthesis because I don't
feel that they it's useful to filter or search for those specific
attributes in the context of a special attack. As an example, GMs
probably don't care to find all monsters that have a special attack of
distraction that has a DC above 10. All special attacks that have a DC
10 or above in general is equally silly.

This again uses my good friend the StringScanner. After finding the
special attacks element, it loads that text into a string scanner. The
scanner skips the "Special Attacks " intro first. Then starts a loop
that ends when the scanner gets to the end of its string.

Inside the loop, the scanner scans for multiple word characters and
spaces. Strings like "gaze" or "spit acid", but not special characters
like "(" or ",". It stores what it scanned over in a buffer. Then is
skips until a comma or the end of the string. If it detects that it's
currently inside a parenthesis, it skips until a close parenthesis,
then scans again for the comma/end-of-line separators.

Finally, it ads the buffer to an array of attack strings. The end of
the method returns that array.