Add concentration parser
Concentration is listed at the start of Spell-like abilities like so:

Spell-Like Abilities (CL 12th; concentration +12)

This is one of the simpler ones. It finds the parent element, looks for
the first digit after the text "concentration " and converts it to an
integer. It returns nil if it doesn't find a parent or if there is no
concentration.