Improve creature parser spec
Since switching to scanning for headers with CR in them, I didn't
account for creatures that have a fractional challenge rating like 1/2
or 1/3.

The parser checks for "CR", an optional space, and multiple digits. Now
it also checks for an optional "/" then digit such as /3 or /2.
Finally, it checks for the end of the line.