Add Ability Scores parser
First, this gets the parent element like any other parser. Then it gets
the parent text like the melee or ranged parsers because of one monk
example class that splits up the ability scores onto two lines.

Once it has the proper text, it removes spaces after commas for parsing
consistency. Then it iterates over each piece text between the commas
"Str 10" "Dex 13", splits them further by the space between them "Str"
"10" then converts them to symbols and integers. In the case that one
of the abilities is not applicable, it converts that ability to nil.

Finally it returns a hash of the abilities.

Notable is the removal of soft hyphens. Soft hyphens are often
invisible characters that nonetheless are a character. One was hiding
in a example paladins (I think) so the parser was getting tripped up.
Stripping them from the text fixed that issue.